// src/lib/shopify/types.ts
export type Image = {
  url: string
  altText: string
  width: number
  height: number
}